requirements.txt
selenium
tensorflow
pillow
matplotlib
numpy